Add async PostgreSQL persister support
Async implementation to persist the state to a postgresql db.

In addition adds methods to postgresql sync persister to make it
consistent. Migrates postgresql to psycopg2 module for consistent
renaming.
