+++ docs/reference/persister.rst
@@ -98,3 +98,8 @@ Currently we support the following, although we highly recommend you contribute 
    :members:
 
    .. automethod:: __init__
+
+.. autoclass:: burr.integrations.persisters.postgresql.AsyncPostgreSQLPersister
+   :members:
+
+   .. automethod:: __init__
