+++ tests/integrations/persisters/test_postgresql.py
@@ -4,7 +4,8 @@ import pickle
 import pytest
 
 from burr.core import state
-from burr.integrations.persisters.postgresql import PostgreSQLPersister
+from burr.integrations.persisters.b_asyncpg import AsyncPostgreSQLPersister
+from burr.integrations.persisters.b_psycopg2 import PostgreSQLPersister
 
 if not os.environ.get("BURR_CI_INTEGRATION_TESTS") == "true":
     pytest.skip("Skipping integration tests", allow_module_level=True)
@@ -22,7 +23,7 @@ def postgresql_persister():
     )
     persister.initialize()
     yield persister
-    persister.connection.close()
+    persister.cleanup()
 
 
 def test_save_and_load_state(postgresql_persister):
@@ -86,3 +87,60 @@ def test_serialization_with_pickle(postgresql_persister):
     data = deserialized_persister.load("pk", "app_id_serde", 1)
 
     assert data["state"].get_all() == {"a": 1, "b": 2}
+
+
+@pytest.fixture
+async def asyncpostgresql_persister():
+    persister = await AsyncPostgreSQLPersister.from_values(
+        db_name="postgres",
+        user="postgres",
+        password="postgres",
+        host="localhost",
+        port=5432,
+        table_name="testtable_async",
+    )
+    await persister.initialize()
+    yield persister
+    await persister.cleanup()
+
+
+async def test_async_pg_fixture(asyncpostgresql_persister):
+    assert await asyncpostgresql_persister.is_initialized()
+
+
+async def test_async_is_initialized_false():
+    persister = await AsyncPostgreSQLPersister.from_values(
+        db_name="postgres",
+        user="postgres",
+        password="postgres",
+        host="localhost",
+        port=5432,
+        table_name="testtable_async2",
+    )
+    assert not await persister.is_initialized()
+
+
+async def test_async_save_and_load_state(asyncpostgresql_persister):
+    await asyncpostgresql_persister.save(
+        "pk", "app_id", 1, "pos", state.State({"a": 1, "b": 2}), "completed"
+    )
+    data = await asyncpostgresql_persister.load("pk", "app_id", 1)
+    print(data)
+    assert data["state"].get_all() == {"a": 1, "b": 2}
+
+
+async def test_async_list_app_ids(asyncpostgresql_persister):
+    await asyncpostgresql_persister.save(
+        "pk", "app_id1", 1, "pos1", state.State({"a": 1}), "completed"
+    )
+    await asyncpostgresql_persister.save(
+        "pk", "app_id2", 2, "pos2", state.State({"b": 2}), "completed"
+    )
+    app_ids = await asyncpostgresql_persister.list_app_ids("pk")
+    assert "app_id1" in app_ids
+    assert "app_id2" in app_ids
+
+
+async def test_async_load_nonexistent_key(asyncpostgresql_persister):
+    state_data = await asyncpostgresql_persister.load("pk", "nonexistent_key")
+    assert state_data is None
